requirements.txt
mysql-connector-python==8.2.0
PyYAML==6.0.1
python-dotenv==1.0.0